requirements.txt
PyPDF2
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
